BookACourt-Backend/requirements.txt
aiohappyeyeballs==2.6.1
aiohttp==3.13.2
aiohttp-retry==2.9.1
aiosignal==1.4.0
asgiref==3.11.0
async-timeout==5.0.1
attrs==25.4.0
certifi==2025.11.12
charset-normalizer==3.4.4
dj-rest-auth==7.0.1
Django==5.2.9
django-allauth==65.13.1
django-cors-headers==4.9.0
django-jazzmin==3.0.1
django-phonenumber-field==8.4.0
djangorestframework==3.16.1
djangorestframework_simplejwt==5.5.1
drf-nested-routers==0.95.0
drf-spectacular==0.29.0
drf-spectacular-sidecar==2025.12.1
frozenlist==1.8.0
idna==3.11
inflection==0.5.1
jsonschema==4.25.1
jsonschema-specifications==2025.9.1
multidict==6.7.0
mysqlclient==2.2.7
phonenumbers==9.0.20
pillow==12.0.0
propcache==0.4.1
PyJWT==2.10.1
pyotp==2.9.0
python-dotenv==1.2.1
PyYAML==6.0.3
referencing==0.37.0
requests==2.32.5
rpds-py==0.30.0
sqlparse==0.5.4
twilio==9.8.8
typing_extensions==4.15.0
uritemplate==4.2.0
urllib3==2.6.0
yarl==1.22.0
